+++ packages/cubejs-schema-compiler/src/compiler/converters/index.ts
@@ -1,0 +1,2 @@
+export * from './CubeSchemaConverter';
+export * from './CubePreAggregationConverter';